+++ epps-labe-food7/src/Routes/Router.js
@@ -1,10 +1,11 @@
 import React from 'react'
 import { BrowserRouter,Switch, Route} from "react-router-dom"
 import CartPage from '../Pages/CartPage/CartPage'
-import DetailsPage from '../Pages/DetailsPage/DetailsPage'
+import FeedPage from '../Pages/FeedPage/FeedPage'
 import HomePage from '../Pages/HomePage/HomePage'
 import LoginPage from '../Pages/LoginPage/LoginPage'
 import ProfilePage from '../Pages/ProfilePage/ProfilePage'
+import DetailsPage from '../Pages/DetailsPage/DetailsPage'
 import RequestsPage from '../Pages/RequestsPage/RequestsPage'
 import SignUpPage from '../Pages/SignUpPage/SignUpPage'
 import ErrorPage from '../Pages/ErrorPage/ErrorPage'
@@ -23,7 +24,6 @@ const Router = () => {
         <Route exact path="/endereco">
           <AdressPage/>
         </Route>
-
         <Route exact path="/">
           <HomePage/>
         </Route>
@@ -31,6 +31,9 @@ const Router = () => {
           <CartPage/>
         </Route>
         <Route exact path="/feed">
+          <FeedPage/>
+        </Route>
+        <Route exact path="/restaurante/:id">
           <DetailsPage/>
         </Route>
         <Route exact path="/pedidos">
